[Sema] fixed bug: set var->defined if found in arguments
