+++ include/Sema/Sema.h
@@ -6,23 +6,14 @@
 #include "AST.h"
 
 namespace Metro::Semantics {
+  struct ScopeContext {
+    AST::Scope*  scope = nullptr;
+    AST::Base* cur_ast = nullptr;
+    size_t  cur_index = 0;
+  };
 
   class Sema {
     using ASTKind = AST::Kind;
-    //using ReturnTypeElement = std::tuple<size_t, AST::Base*, ValueType>;
-
-    struct ReturnTypeElement {
-      bool    is_analyzed = false;
-      size_t  index;
-      AST::Base*  ast;
-      ValueType   type;
-    };
-
-    struct ScopeContext {
-      AST::Scope*  scope = nullptr;
-      AST::Base* cur_ast = nullptr;
-      size_t  cur_index = 0;
-    };
 
   public:
 
@@ -33,35 +24,40 @@ namespace Metro::Semantics {
     ValueType walk(AST::Base* ast);
 
 
+    //
     // 関数の戻り値の型を解析する
     void analyze_func_return_type(ValueType& out, AST::Function* ast);
 
+    //
     // return 式を探して out に追加する
     void find_return(std::vector<AST::Base*>& out, AST::Base* ast);
 
+    //
     // 全ての last-expr を out に追加する
     // 注意: return は含まれません
     void get_lastvalues(std::vector<AST::Base*>& out, AST::Base* ast);
 
+    //
     // 式の結果になりうる全ての式を out に追加する
     //  last-expr と return どっちも追加される
     void get_lastval_full(std::vector<AST::Base*>& out, AST::Base* ast);
 
+    //
     // 式の中に name と同名の関数呼び出しが含まれているかどうか確認
     //  !!! ast には AST::Expr を渡すこと !!!
     //  !!! 'ast' must AST::Expr !!!
     bool contains_callfunc_in_expr(std::string_view name, AST::Base* ast);
 
-
+    //
+    // create object from token
     static Object* create_obj(Token* token);
 
   private:
 
     AST::Scope* root;
 
-    // informations of function that current walking ; cfn = current-function //
+    // current walking function
     AST::Function* cfn_ast;
-    std::vector<ReturnTypeElement>* cfn_ret_list;
 
     // all functions in root
     std::vector<AST::Function*> functions;
@@ -73,5 +69,4 @@ namespace Metro::Semantics {
     std::map<AST::Base*, ValueType> walked;
 
   };
-
 }
